Separate depinject
